Unprotect the services endpoint
